Refactor: Reorganize core types into separate files and Core folder
- Split Core.swift into individual files for better maintainability
- Create Core/ directory to organize core components:
  - CodeElement.swift - CodeElement protocol
  - CodeToken.swift - CodeToken protocol
  - CodeTokenizer.swift - CodeTokenizer protocol
  - CodeTokenConsumer.swift - CodeTokenConsumer protocol
  - CodeNode.swift - CodeNode class with all tree operations
  - CodeError.swift - CodeError struct
  - CodeContext.swift - CodeContext struct
  - CodeLanguage.swift - CodeLanguage protocol
  - CodeParser.swift - Moved from root to Core folder
- Fix Package.swift deprecation warning by replacing swiftLanguageVersions with swiftLanguageModes
- All tests continue to pass after refactoring
