+++ Sources/SwiftParser/Core/CodeElement.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+public protocol CodeElement {}
