+++ Sources/SwiftParser/Core/CodeLanguage.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public protocol CodeLanguage {
+    var tokenizer: CodeTokenizer { get }
+    var consumers: [CodeTokenConsumer] { get }
+    var rootElement: any CodeElement { get }
+}
