+++ Sources/SwiftParser/Core/CodeTokenizer.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol CodeTokenizer {
+    func tokenize(_ input: String) -> [any CodeToken]
+}
